requirements.txt
# Core Dependencies
langchain
langchain-openai
# Utilities
python-dotenv